Checking In 3/2/2018
Attached is a list of stuff that we have done and need to be done, I will try to keep this as up to date as possible. In the file under the "Things To Do" column, I have assigned duties to each person. Some of the tasks have been deemed important and are highlighted. I'm hoping that the highlighted items can be completed by the next time we have class, Friday the 9th.